requirements.txt
PyMuPDF==1.23.8
python-docx==1.0.1
nltk==3.8.1
scikit-learn==1.4.0
sentence-transformers==2.5.1
spacy==3.7.4
keybert==0.7.0
numpy==1.26.4
pandas==2.2.1
torch>=2.0.0
transformers>=4.30.0 